pyproject.toml
[project]
name = "mike-flores-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "playwright>=1.50.0",
]
